+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -90,6 +90,10 @@ Notes](../../RELEASENOTES.md).
 * Fix `ArgumentException` if `OTEL_SDK_DISABLED=true`.
   ([#7273](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7273))
 
+* GZip compress scrape endpoint responses when `Accept-Encoding: gzip` is
+  specified by the HTTP request headers.
+  ([#7274](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7274))
+
 ## 1.15.3-beta.1
 
 Released 2026-Apr-21